Addressed all comments from second code reviewer
